pyproject.toml
[tool.poetry]
name = "ska-low-csp-testware"
version = "0.1.0"
description = ""
authors = ["Sander Ploegsma <sander.ploegsma@topic.nl>"]
license = "BSD-3-Clause"
readme = "README.md"

[tool.poetry.scripts]
pcap_file_monitor = 'ska_low_csp_testware.pcap_file_monitor:main'

[[tool.poetry.source]]
name = "default"
url = "https://pypi.org/simple"

[[tool.poetry.source]]
name = "ska-nexus"
url = "https://artefact.skao.int/repository/pypi-internal/simple"

[tool.poetry.dependencies]
python = "^3.10"
pytango = "^9.5.0"
asyncio = "^3.4.3"
spead2 = "^4.3.1"
pandas = "^2.2.1"
ska-control-model = "^0.3.4"
typing-extensions = "^4.10.0"
watchfiles = "^0.21.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pytest-cov = "^4.1.0"
ska-tango-testing = "^0.6.1"
pytest-forked = "^1.6.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
flake8 = "^6.1.0"
isort = "^5.13.2"
pylint = "^3.0.3"
pylint-junit = "^0.3.4"

[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
ska-ser-sphinx-theme = "^0.1.2"
sphinx-autoapi = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.3"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 127

[tool.flake8]
max-line-length = 127

[tool.isort]
profile = "black"
line_length = 127

[tool.pylint.format]
max-line-length = 127

[tool.pylint.messages_control]
disable = ["duplicate-code"]
